Make withAuthenticatedUser pass user authentication status to other components
The goal of this change is to eliminate the repeated subscriptions
for authentication status to one place.

- Replaced isAuthenticated with authenticationStatus observable that
is able to pass one of the 3 values (strings):

in_progress - fetching user data.
not_authenticated - user is not authenticated.
authenticated - user is authenticated.

- Removed all subscriptions to isAuthenticated,
withAuthenticatedUser component now will pass authenticationStatus
property to other components